+++ packages/jsenv-lighthouse-impact/src/internal/formatLighthouseReportForLog.js
@@ -1,7 +1,0 @@
-export const formatLighthouseReportForLog = (lighthouseReport) => {
-  const scores = {}
-  Object.keys(lighthouseReport.categories).forEach((name) => {
-    scores[name] = lighthouseReport.categories[name].score
-  })
-  return JSON.stringify(scores, null, "  ")
-}