+++ packages/jsenv-lighthouse-impact/src/main.js
@@ -1,4 +1,4 @@
 export { readGitHubWorkflowEnv } from "@jsenv/github-pull-request-impact"
 
-export { generateLighthouseReport } from "./generateLighthouseReport.js"
-export { reportLighthouseImpact } from "./reportLighthouseImpact.js"
+export { runLighthouseOnPlaywrightPage } from "./run_lighthouse_on_playwright_page.js"
+export { reportLighthouseImpactInGithubPullRequest } from "./report_lighthouse_impact_in_github_pr.js"